fix: resolve all flake8, mypy, and cSpell errors across codebase
Zero flake8 errors, zero mypy errors, 27/27 tests passing.
Added comprehensive cSpell dictionary for financial domain terms,
type annotations for dict[str, Any] patterns, and type: ignore
comments for pandas-stubs limitations.

+++ grri_mac/pillars/private_credit_pca.py
@@ -107,7 +107,7 @@ class RollingPCADecorrelator:
 
         if min_len < self.min_observations:
             return PCADecorrelationResult(
-                decorrelated_score=np.mean(bdc_returns[-20:])
+                decorrelated_score=float(np.mean(bdc_returns[-20:]))
                 if len(bdc_returns) >= 20
                 else 0.5,
                 raw_score=0.5,
